+++ src/main/java/com/drakmyth/minecraft/manufactory/gui/BallMillGui.java
@@ -7,7 +7,7 @@ package com.drakmyth.minecraft.manufactory.gui;
 
 import com.drakmyth.minecraft.manufactory.LogMarkers;
 import com.drakmyth.minecraft.manufactory.Reference;
-import com.drakmyth.minecraft.manufactory.menus.BallMillContainer;
+import com.drakmyth.minecraft.manufactory.menus.BallMillMenu;
 import com.mojang.blaze3d.systems.RenderSystem;
 import com.mojang.blaze3d.vertex.PoseStack;
 
@@ -20,11 +20,11 @@ import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.network.chat.Component;
 
-public class BallMillGui extends AbstractContainerScreen<BallMillContainer> {
+public class BallMillGui extends AbstractContainerScreen<BallMillMenu> {
     private static final ResourceLocation GUI_TEXTURE = new ResourceLocation(Reference.MOD_ID, "textures/gui/ball_mill.png");
     private static final Logger LOGGER = LogManager.getLogger();
 
-    public BallMillGui(BallMillContainer screenContainer, Inventory inv, Component name) {
+    public BallMillGui(BallMillMenu screenContainer, Inventory inv, Component name) {
         super(screenContainer, inv, name);
     }
