+++ src/main/java/com/drakmyth/minecraft/manufactory/menus/BallMillMenu.java
@@ -8,9 +8,6 @@ package com.drakmyth.minecraft.manufactory.menus;
 import com.drakmyth.minecraft.manufactory.LogMarkers;
 import com.drakmyth.minecraft.manufactory.init.ModBlocks;
 import com.drakmyth.minecraft.manufactory.init.ModContainerTypes;
-import com.drakmyth.minecraft.manufactory.items.upgrades.IMillingBallUpgrade;
-import com.drakmyth.minecraft.manufactory.items.upgrades.IMotorUpgrade;
-import com.drakmyth.minecraft.manufactory.items.upgrades.IPowerUpgrade;
 import com.drakmyth.minecraft.manufactory.tileentities.BallMillTileEntity;
 
 import org.apache.logging.log4j.LogManager;
@@ -30,56 +27,37 @@ import net.minecraftforge.items.ItemStackHandler;
 import net.minecraftforge.items.SlotItemHandler;
 import net.minecraftforge.items.wrapper.InvWrapper;
 
-public class BallMillUpgradeContainer extends AbstractContainerMenu {
+public class BallMillMenu extends AbstractContainerMenu {
     private static final Logger LOGGER = LogManager.getLogger();
-    public final ItemStackHandler upgradeInventory;
+
+    public final ItemStackHandler ballMillInventory;
     private final ContainerLevelAccess posCallable;
     private final BallMillTileEntity tileEntity;
 
-    public BallMillUpgradeContainer(int windowId, Inventory playerInventory, FriendlyByteBuf data) {
+    public BallMillMenu(int windowId, Inventory playerInventory, FriendlyByteBuf data) {
         this(windowId, new InvWrapper(playerInventory), playerInventory.player, data.readBlockPos());
     }
 
-    public BallMillUpgradeContainer(int windowId, IItemHandler playerInventory, Player player, BlockPos pos) {
-        super(ModContainerTypes.BALL_MILL_UPGRADE.get(), windowId);
-        LOGGER.debug(LogMarkers.CONTAINER, "Initializing BallMillUpgradeContainer...");
+    public BallMillMenu(int windowId, IItemHandler playerInventory, Player player, BlockPos pos) {
+        super(ModContainerTypes.BALL_MILL.get(), windowId);
+        LOGGER.debug(LogMarkers.CONTAINER, "Initializing BallMillContainer...");
         Level world = player.getCommandSenderWorld();
         posCallable = ContainerLevelAccess.create(world, pos);
         tileEntity = (BallMillTileEntity)world.getBlockEntity(pos);
-        upgradeInventory = tileEntity.getUpgradeInventory();
+        ballMillInventory = tileEntity.getInventory();
 
-        // Ball Mill Slots
-        // Milling Ball Slot
-        this.addSlot(new SlotItemHandler(upgradeInventory, 0, 44, 36) {
+        // Grinder Slots
+        // Input Slot
+        this.addSlot(new SlotItemHandler(ballMillInventory, 0, 56, 35));
+        LOGGER.debug(LogMarkers.CONTAINER, "Input slot added with index 0");
+        // Output Slot
+        this.addSlot(new SlotItemHandler(ballMillInventory, 1, 116, 35) {
             @Override
             public boolean mayPlace(ItemStack stack) {
-                return stack.getItem() instanceof IMillingBallUpgrade;
+                return false;
             }
         });
-        LOGGER.debug(LogMarkers.CONTAINER, "Milling ball slot added with index 0");
-        // Motor Slot
-        this.addSlot(new SlotItemHandler(upgradeInventory, 1, 80, 36) {
-            @Override
-            public boolean mayPlace(ItemStack stack) {
-                return stack.getItem() instanceof IMotorUpgrade;
-            }
-        });
-        LOGGER.debug(LogMarkers.CONTAINER, "Motor slot added with index 1");
-        // Power Slot
-        this.addSlot(new SlotItemHandler(upgradeInventory, 2, 80, 58) {
-            @Override
-            public boolean mayPlace(ItemStack stack) {
-                return stack.getItem() instanceof IPowerUpgrade;
-            }
-
-            @Override
-            public void setChanged() {
-                super.setChanged();
-                LOGGER.debug(LogMarkers.CONTAINER, "Power slot contents changed. Notifying neighbors...");
-                tileEntity.getBlockState().updateNeighbourShapes(world, pos, 3);
-            }
-        });
-        LOGGER.debug(LogMarkers.CONTAINER, "Power slot added with index 2");
+        LOGGER.debug(LogMarkers.CONTAINER, "Output slot added with index 1");
 
         // Player Inventory
         for (int j = 0; j < 3; j++) {
@@ -104,14 +82,14 @@ public class BallMillUpgradeContainer extends AbstractContainerMenu {
         if (slot != null && slot.hasItem()) {
             ItemStack itemstack1 = slot.getItem();
             itemstack = itemstack1.copy();
-            if (index < upgradeInventory.getSlots()) { // transfer from ball mill to inventory
-                LOGGER.debug(LogMarkers.CONTAINER, "Transferring stack from ball mill upgrade slot {} to player inventory...", index);
-                if (!this.moveItemStackTo(itemstack1, upgradeInventory.getSlots(), this.slots.size(), false)) {
+            if (index < ballMillInventory.getSlots()) { // transfer from ball mill to inventory
+                LOGGER.debug(LogMarkers.CONTAINER, "Transferring stack from ball mill slot {} to player inventory...", index);
+                if (!this.moveItemStackTo(itemstack1, ballMillInventory.getSlots(), this.slots.size(), false)) {
                     LOGGER.debug(LogMarkers.CONTAINER, "Transfer failed because player inventory is full");
                     return ItemStack.EMPTY;
                 }
-            } else if (!this.moveItemStackTo(itemstack1, 0, upgradeInventory.getSlots(), false)) { // transfer from inventory to ball mill
-                LOGGER.debug(LogMarkers.CONTAINER, "Transfer of stack from player inventory slot {} to ball mill upgrade inventory failed because upgrade inputs are full", index);
+            } else if (!this.moveItemStackTo(itemstack1, 0, 1, false)) { // transfer from inventory to ball mill
+                LOGGER.debug(LogMarkers.CONTAINER, "Transfer of stack from player inventory slot {} to ball mill failed because ball mill input is full", index);
                 return ItemStack.EMPTY;
             }
