+++ src/main/java/com/drakmyth/minecraft/manufactory/menus/GrinderMenu.java
@@ -8,9 +8,6 @@ package com.drakmyth.minecraft.manufactory.menus;
 import com.drakmyth.minecraft.manufactory.LogMarkers;
 import com.drakmyth.minecraft.manufactory.init.ModBlocks;
 import com.drakmyth.minecraft.manufactory.init.ModContainerTypes;
-import com.drakmyth.minecraft.manufactory.items.upgrades.IGrinderWheelUpgrade;
-import com.drakmyth.minecraft.manufactory.items.upgrades.IMotorUpgrade;
-import com.drakmyth.minecraft.manufactory.items.upgrades.IPowerUpgrade;
 import com.drakmyth.minecraft.manufactory.tileentities.GrinderTileEntity;
 
 import org.apache.logging.log4j.LogManager;
@@ -30,65 +27,37 @@ import net.minecraftforge.items.ItemStackHandler;
 import net.minecraftforge.items.SlotItemHandler;
 import net.minecraftforge.items.wrapper.InvWrapper;
 
-public class GrinderUpgradeContainer extends AbstractContainerMenu {
+public class GrinderMenu extends AbstractContainerMenu {
     private static final Logger LOGGER = LogManager.getLogger();
 
-    public final ItemStackHandler upgradeInventory;
+    public final ItemStackHandler grinderInventory;
     private final ContainerLevelAccess posCallable;
     private final GrinderTileEntity tileEntity;
 
-    public GrinderUpgradeContainer(int windowId, Inventory playerInventory, FriendlyByteBuf data) {
+    public GrinderMenu(int windowId, Inventory playerInventory, FriendlyByteBuf data) {
         this(windowId, new InvWrapper(playerInventory), playerInventory.player, data.readBlockPos());
     }
 
-    public GrinderUpgradeContainer(int windowId, IItemHandler playerInventory, Player player, BlockPos pos) {
-        super(ModContainerTypes.GRINDER_UPGRADE.get(), windowId);
-        LOGGER.debug(LogMarkers.CONTAINER, "Initializing GrinderUpgradeContainer...");
+    public GrinderMenu(int windowId, IItemHandler playerInventory, Player player, BlockPos pos) {
+        super(ModContainerTypes.GRINDER.get(), windowId);
+        LOGGER.debug(LogMarkers.CONTAINER, "Initializing GrinderContainer...");
         Level world = player.getCommandSenderWorld();
         posCallable = ContainerLevelAccess.create(world, pos);
         tileEntity = (GrinderTileEntity)world.getBlockEntity(pos);
-        upgradeInventory = tileEntity.getUpgradeInventory();
+        grinderInventory = tileEntity.getInventory();
 
         // Grinder Slots
-        // Wheel Slot 1
-        this.addSlot(new SlotItemHandler(upgradeInventory, 0, 62, 14) {
+        // Input Slot
+        this.addSlot(new SlotItemHandler(grinderInventory, 0, 56, 35));
+        LOGGER.debug(LogMarkers.CONTAINER, "Input slot added with index 0");
+        // Output Slot
+        this.addSlot(new SlotItemHandler(grinderInventory, 1, 116, 35) {
             @Override
             public boolean mayPlace(ItemStack stack) {
-                return stack.getItem() instanceof IGrinderWheelUpgrade;
+                return false;
             }
         });
-        LOGGER.debug(LogMarkers.CONTAINER, "Grinder wheel slot 1 added with index 0");
-        // Wheel Slot 2
-        this.addSlot(new SlotItemHandler(upgradeInventory, 1, 98, 14) {
-            @Override
-            public boolean mayPlace(ItemStack stack) {
-                return stack.getItem() instanceof IGrinderWheelUpgrade;
-            }
-        });
-        LOGGER.debug(LogMarkers.CONTAINER, "Grinder wheel slot 2 added with index 1");
-        // Motor Slot
-        this.addSlot(new SlotItemHandler(upgradeInventory, 2, 80, 36) {
-            @Override
-            public boolean mayPlace(ItemStack stack) {
-                return stack.getItem() instanceof IMotorUpgrade;
-            }
-        });
-        LOGGER.debug(LogMarkers.CONTAINER, "Motor slot added with index 2");
-        // Power Slot
-        this.addSlot(new SlotItemHandler(upgradeInventory, 3, 80, 58) {
-            @Override
-            public boolean mayPlace(ItemStack stack) {
-                return stack.getItem() instanceof IPowerUpgrade;
-            }
-
-            @Override
-            public void setChanged() {
-                super.setChanged();
-                LOGGER.debug(LogMarkers.CONTAINER, "Power slot contents changed. Notifying neighbors...");
-                tileEntity.getBlockState().updateNeighbourShapes(world, pos, 3);
-            }
-        });
-        LOGGER.debug(LogMarkers.CONTAINER, "Power slot added with index 3");
+        LOGGER.debug(LogMarkers.CONTAINER, "Output slot added with index 1");
 
         // Player Inventory
         for (int j = 0; j < 3; j++) {
@@ -113,14 +82,14 @@ public class GrinderUpgradeContainer extends AbstractContainerMenu {
         if (slot != null && slot.hasItem()) {
             ItemStack itemstack1 = slot.getItem();
             itemstack = itemstack1.copy();
-            if (index < upgradeInventory.getSlots()) { // transfer from grinder to inventory
-                LOGGER.debug(LogMarkers.CONTAINER, "Transferring stack from grinder upgrade slot {} to player inventory...", index);
-                if (!this.moveItemStackTo(itemstack1, upgradeInventory.getSlots(), this.slots.size(), false)) {
+            if (index < grinderInventory.getSlots()) { // transfer from grinder to inventory
+                LOGGER.debug(LogMarkers.CONTAINER, "Transferring stack from grinder slot {} to player inventory...", index);
+                if (!this.moveItemStackTo(itemstack1, grinderInventory.getSlots(), this.slots.size(), false)) {
                     LOGGER.debug(LogMarkers.CONTAINER, "Transfer failed because player inventory is full");
                     return ItemStack.EMPTY;
                 }
             } else if (!this.moveItemStackTo(itemstack1, 0, 1, false)) { // transfer from inventory to grinder
-                LOGGER.debug(LogMarkers.CONTAINER, "Transfer of stack from player inventory slot {} to grinder upgrade inventory failed because upgrade inputs are full", index);
+                LOGGER.debug(LogMarkers.CONTAINER, "Transfer of stack from player inventory slot {} to grinder failed because grinder input is full", index);
                 return ItemStack.EMPTY;
             }
