+++ src/main/java/com/drakmyth/minecraft/manufactory/menus/GrinderUpgradeMenu.java
@@ -8,7 +8,10 @@ package com.drakmyth.minecraft.manufactory.menus;
 import com.drakmyth.minecraft.manufactory.LogMarkers;
 import com.drakmyth.minecraft.manufactory.init.ModBlocks;
 import com.drakmyth.minecraft.manufactory.init.ModContainerTypes;
-import com.drakmyth.minecraft.manufactory.tileentities.BallMillTileEntity;
+import com.drakmyth.minecraft.manufactory.items.upgrades.IGrinderWheelUpgrade;
+import com.drakmyth.minecraft.manufactory.items.upgrades.IMotorUpgrade;
+import com.drakmyth.minecraft.manufactory.items.upgrades.IPowerUpgrade;
+import com.drakmyth.minecraft.manufactory.tileentities.GrinderTileEntity;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -27,37 +30,65 @@ import net.minecraftforge.items.ItemStackHandler;
 import net.minecraftforge.items.SlotItemHandler;
 import net.minecraftforge.items.wrapper.InvWrapper;
 
-public class BallMillContainer extends AbstractContainerMenu {
+public class GrinderUpgradeMenu extends AbstractContainerMenu {
     private static final Logger LOGGER = LogManager.getLogger();
 
-    public final ItemStackHandler ballMillInventory;
+    public final ItemStackHandler upgradeInventory;
     private final ContainerLevelAccess posCallable;
-    private final BallMillTileEntity tileEntity;
+    private final GrinderTileEntity tileEntity;
 
-    public BallMillContainer(int windowId, Inventory playerInventory, FriendlyByteBuf data) {
+    public GrinderUpgradeMenu(int windowId, Inventory playerInventory, FriendlyByteBuf data) {
         this(windowId, new InvWrapper(playerInventory), playerInventory.player, data.readBlockPos());
     }
 
-    public BallMillContainer(int windowId, IItemHandler playerInventory, Player player, BlockPos pos) {
-        super(ModContainerTypes.BALL_MILL.get(), windowId);
-        LOGGER.debug(LogMarkers.CONTAINER, "Initializing BallMillContainer...");
+    public GrinderUpgradeMenu(int windowId, IItemHandler playerInventory, Player player, BlockPos pos) {
+        super(ModContainerTypes.GRINDER_UPGRADE.get(), windowId);
+        LOGGER.debug(LogMarkers.CONTAINER, "Initializing GrinderUpgradeContainer...");
         Level world = player.getCommandSenderWorld();
         posCallable = ContainerLevelAccess.create(world, pos);
-        tileEntity = (BallMillTileEntity)world.getBlockEntity(pos);
-        ballMillInventory = tileEntity.getInventory();
+        tileEntity = (GrinderTileEntity)world.getBlockEntity(pos);
+        upgradeInventory = tileEntity.getUpgradeInventory();
 
         // Grinder Slots
-        // Input Slot
-        this.addSlot(new SlotItemHandler(ballMillInventory, 0, 56, 35));
-        LOGGER.debug(LogMarkers.CONTAINER, "Input slot added with index 0");
-        // Output Slot
-        this.addSlot(new SlotItemHandler(ballMillInventory, 1, 116, 35) {
+        // Wheel Slot 1
+        this.addSlot(new SlotItemHandler(upgradeInventory, 0, 62, 14) {
             @Override
             public boolean mayPlace(ItemStack stack) {
-                return false;
+                return stack.getItem() instanceof IGrinderWheelUpgrade;
             }
         });
-        LOGGER.debug(LogMarkers.CONTAINER, "Output slot added with index 1");
+        LOGGER.debug(LogMarkers.CONTAINER, "Grinder wheel slot 1 added with index 0");
+        // Wheel Slot 2
+        this.addSlot(new SlotItemHandler(upgradeInventory, 1, 98, 14) {
+            @Override
+            public boolean mayPlace(ItemStack stack) {
+                return stack.getItem() instanceof IGrinderWheelUpgrade;
+            }
+        });
+        LOGGER.debug(LogMarkers.CONTAINER, "Grinder wheel slot 2 added with index 1");
+        // Motor Slot
+        this.addSlot(new SlotItemHandler(upgradeInventory, 2, 80, 36) {
+            @Override
+            public boolean mayPlace(ItemStack stack) {
+                return stack.getItem() instanceof IMotorUpgrade;
+            }
+        });
+        LOGGER.debug(LogMarkers.CONTAINER, "Motor slot added with index 2");
+        // Power Slot
+        this.addSlot(new SlotItemHandler(upgradeInventory, 3, 80, 58) {
+            @Override
+            public boolean mayPlace(ItemStack stack) {
+                return stack.getItem() instanceof IPowerUpgrade;
+            }
+
+            @Override
+            public void setChanged() {
+                super.setChanged();
+                LOGGER.debug(LogMarkers.CONTAINER, "Power slot contents changed. Notifying neighbors...");
+                tileEntity.getBlockState().updateNeighbourShapes(world, pos, 3);
+            }
+        });
+        LOGGER.debug(LogMarkers.CONTAINER, "Power slot added with index 3");
 
         // Player Inventory
         for (int j = 0; j < 3; j++) {
@@ -82,14 +113,14 @@ public class BallMillContainer extends AbstractContainerMenu {
         if (slot != null && slot.hasItem()) {
             ItemStack itemstack1 = slot.getItem();
             itemstack = itemstack1.copy();
-            if (index < ballMillInventory.getSlots()) { // transfer from ball mill to inventory
-                LOGGER.debug(LogMarkers.CONTAINER, "Transferring stack from ball mill slot {} to player inventory...", index);
-                if (!this.moveItemStackTo(itemstack1, ballMillInventory.getSlots(), this.slots.size(), false)) {
+            if (index < upgradeInventory.getSlots()) { // transfer from grinder to inventory
+                LOGGER.debug(LogMarkers.CONTAINER, "Transferring stack from grinder upgrade slot {} to player inventory...", index);
+                if (!this.moveItemStackTo(itemstack1, upgradeInventory.getSlots(), this.slots.size(), false)) {
                     LOGGER.debug(LogMarkers.CONTAINER, "Transfer failed because player inventory is full");
                     return ItemStack.EMPTY;
                 }
-            } else if (!this.moveItemStackTo(itemstack1, 0, 1, false)) { // transfer from inventory to ball mill
-                LOGGER.debug(LogMarkers.CONTAINER, "Transfer of stack from player inventory slot {} to ball mill failed because ball mill input is full", index);
+            } else if (!this.moveItemStackTo(itemstack1, 0, 1, false)) { // transfer from inventory to grinder
+                LOGGER.debug(LogMarkers.CONTAINER, "Transfer of stack from player inventory slot {} to grinder upgrade inventory failed because upgrade inputs are full", index);
                 return ItemStack.EMPTY;
             }
 
@@ -118,6 +149,6 @@ public class BallMillContainer extends AbstractContainerMenu {
 
     @Override
     public boolean stillValid(Player playerIn) {
-        return stillValid(posCallable, playerIn, ModBlocks.BALL_MILL.get());
+        return stillValid(posCallable, playerIn, ModBlocks.GRINDER.get());
     }
 }
